add the rst instruction remove the files under script and merge the functions into edl/docker test=develop
