+++ docker/paddle_k8s
@@ -62,27 +62,33 @@ check_trainer_ret() {
 start_fluid_process() {
   pserver_label="paddle-job-pserver=${PADDLE_JOB_NAME}"
   trainer_label="paddle-job=${PADDLE_JOB_NAME}"
+  hostname=${HOSTNAME}
   task_index=""
 
   if [ "${PADDLE_TRAINING_ROLE}" == "TRAINER" ] || [ "${PADDLE_TRAINING_ROLE}" == "PSERVER" ]; then
-    stdbuf -oL python /root/k8s_tools.py wait_pods_running ${pserver_label} ${PADDLE_PSERVERS}
+    stdbuf -oL python /root/k8s_tools.py wait_pods_running ${pserver_label} ${PADDLE_PSERVERS_NUM}
   fi
 
   if [ "${PADDLE_TRAINING_ROLE}" == "TRAINER" ] || [ "${PADDLE_TRAINING_ROLE}" == "WORKER" ]; then
-    stdbuf -oL python /root/k8s_tools.py wait_pods_running ${trainer_label} ${PADDLE_TRAINERS}
+    stdbuf -oL python /root/k8s_tools.py wait_pods_running ${trainer_label} ${PADDLE_TRAINERS_NUM}
   fi
 
-  export PADDLE_PSERVER_IPS=$(python /root/k8s_tools.py fetch_ips ${pserver_label} ${PADDLE_INIT_PORT})
-  export PADDLE_TRAINER_IPS=$(python /root/k8s_tools.py fetch_ips ${trainer_label})
+  export PADDLE_PSERVERS=$(python /root/k8s_tools.py fetch_endpoints ${pserver_label} ${PADDLE_PORT})
+  # export PADDLE_TRAINER_IPS=$(python /root/k8s_tools.py fetch_ips ${trainer_label})
 
   if [ "${PADDLE_TRAINING_ROLE}" == "TRAINER" ] || [ "${PADDLE_TRAINING_ROLE}" == "WORKER" ]; then
     check_failed_cnt 1
     task_index=$(python /root/k8s_tools.py fetch_id ${trainer_label})
+    # download training materials for trainer task_index=0-9
+    file_index=$(( 10$task_index%10 ))
+    address="https://paddle-ctr-data.bj.bcebos.com/dac${file_index}.tar.gz"
+    cd /workspace/ctr/data/raw && curl -o dac.tar.gz ${address} && tar zxf dac.tar.gz && rm dac.tar.gz
   else
     task_index=$(python /root/k8s_tools.py fetch_id ${pserver_label})
   fi
   
   export PADDLE_TRAINER_ID=${task_index}
+  export PADDLE_PSERVER_ID=${task_index}
 
   stdbuf -oL sh -c "${ENTRY}"
   check_trainer_ret $?
