+++ example/ctr/ps-train/trainer.yaml
@@ -18,7 +18,6 @@ spec:
         - {name: TOPOLOGY, value: ''}
         - {name: TRAINER_PACKAGE, value: /workspace}
         - {name: PADDLE_INIT_NICS, value: eth2}
-        #- {name: FLAGS_fraction_of_gpu_memory_to_use, value: "0.92"}
         - name: NAMESPACE
           valueFrom:
             fieldRef: {fieldPath: metadata.namespace}
